+++ src/Helper/GetGracePeriod.php
@@ -2,8 +2,8 @@
 
 namespace umulmrum\Holiday\Helper;
 
-use umulmrum\Holiday\Calculator\HolidayCalculator;
-use umulmrum\Holiday\Calculator\HolidayCalculatorInterface;
+use umulmrum\Holiday\HolidayCalculator;
+use umulmrum\Holiday\HolidayCalculatorInterface;
 use umulmrum\Holiday\Provider\HolidayProviderInterface;
 
 final class GetGracePeriod