+++ tests/Helper/GetNoWorkDaysForTimeSpanTest.php
@@ -11,7 +11,7 @@
 
 namespace umulmrum\Holiday\Test\Helper;
 
-use umulmrum\Holiday\Calculator\HolidayCalculator;
+use umulmrum\Holiday\HolidayCalculator;
 use umulmrum\Holiday\Helper\GetNoWorkDaysForTimeSpan;
 use umulmrum\Holiday\Model\HolidayList;
 use umulmrum\Holiday\Provider\Germany\BadenWuerttemberg;